requirements.txt
requests
python-decouple
numpy
